requirements.txt
flask
python-dotenv
firebase-admin
google-generativeai
google-analytics-data
google-analytics-admin
google-auth-oauthlib
requests
gunicorn
waitress
whitenoise
flask-compress
pytrends
markdown
resend
numpy
APScheduler
gspread
google-auth